adaptions for optimization tests, keras wrapper and rename save_memory -> decrease_memory
