+++ photonai/requirements.txt
@@ -16,10 +16,3 @@ dask
 distributed
 scikit-optimize
 pytest
-
-# bcolz
-# fabolas
-# george
-# emcee
-# you need FORTRAN compiler for the DIRECT package: sudo apt-get install gfortran
-# DIRECT
