+++ test/optimization_tests/random_search_tests/test_random_search.py
@@ -2,7 +2,7 @@ import time
 
 from photonai.base import PipelineElement
 from photonai.optimization import RandomSearchOptimizer, IntegerRange
-from ..grid_search.test_grid_search import GridSearchOptimizerTest
+from ..grid_search_tests.test_grid_search import GridSearchOptimizerTest
 
 
 class RandomSearchOptimizerTest(GridSearchOptimizerTest):
