+++ test/processing_tests/test_metrics.py
@@ -1,6 +1,7 @@
 import unittest
 import types
 import numpy as np
+import warnings
 
 from photonai.processing.metrics import Scorer, spearman_correlation, specificity, sensitivity, one_hot_to_binary, \
     pearson_correlation, balanced_accuracy, categorical_accuracy_score, variance_explained_score
@@ -64,6 +65,10 @@ class ScorerTest(unittest.TestCase):
         with self.assertRaises(ValueError):
             Scorer.register_custom_metric(None)
 
+        with warnings.catch_warnings(record=True) as w:
+            Scorer.register_custom_metric(('a_custom_metric', custom_metric))
+            assert any("is ambiguous. Please specify metric" in s for s in [e.message.args[0] for e in w])
+
     def test_keras_metric(self):
         try:
             from keras.metrics import MeanAbsoluteError
@@ -106,11 +111,3 @@ class ScorerTest(unittest.TestCase):
                               np.concatenate((np.zeros((100,)), np.ones((100,))))), axis=1)
         binarized_multidim = one_hot_to_binary(y_one_hot)
         self.assertTrue(np.array_equal(binarized_multidim, np.concatenate((np.zeros((100,)), np.ones((100, ))))))
-
-
-
-
-
-
-
-
